requirements.txt
bs4~=0.0.2
JPype1~=1.5.1
numpy~=2.2.0
opencv-python~=4.10.0.84
pandas~=2.2.3
pillow~=11.0.0
pytesseract~=0.3.13
requests~=2.32.3
scikit-learn~=1.5.2
sentence-transformers~=3.3.1
tabula-py~=2.10.0
urllib3~=2.2.3
